variabele prijs aanpassen

+++ src/collections/hard/h1/Smartphones.java
@@ -8,7 +8,7 @@ public class Smartphones extends Products{
         return screenSmartphones;
     }
 
-    public Smartphones(String brandname, int modelnumber, float price, ScreenSmartphones screenSmartphones) {
+    public Smartphones(String brandname, int modelnumber, int price, ScreenSmartphones screenSmartphones) {
         super(brandname, modelnumber, price);
         this.screenSmartphones = screenSmartphones;
     }
